Read and provide state when necessary
Adds a STANDALONE env flag. When it's set to True (default) then the ai-traineree is served but no agent is created. To create an agent one should do POST `/agent` request.
When STANDALONE=false (or '0'), the agent will be created on initiation with the state taken from the database. It's expected to have AGENT_ID as the ENV param.